Get xml element with songs and loop through them.
The songs in an iTunes XML library are in the root plist element, and then
under two dict elements (/plist/dict/dict/ as an XPath, I believe).  We get
this node, and then iterate through all of its children (by getting its first
child and then repeatedly getting the next sibling).  For each song, we
trivially print out the song number and the text in the first "key" node.