For each song, print out the title and artist.
When looping through all songs in the library, loop through all "key" child
elements.  For "Name" and "Artist" elements, get the next sibling, which is
the value of that key.  Print these values out when they are found, using the
assumption that Name keys appear before "Artist" keys.  Also include the
number of the song.