Added header and definition for class Song.
The Song class represents a single song, with some metadata about that song.
For now, this metadata is what can be extracted from an iTunes library xml
file.  There is a helper function, create_from_iTunes_node, that does exactly
that: it uses the data in an object representing the "dict" tag in the XML
library file to create a Song object.