Only print parsed songs if print setting is true.
Added a setting, "print", to indicate whether the songs parsed from the
library file should be printed.  If true, they are printed, as before.  If
false, none of the song information is printed.

For now, the print setting is simply a local variable and cannot be changed
from outside the program.