Added a default constructor to set default values.
There appeared to be a bug where songs' ratings were the greater of the value
found in the iTunes library file and the highest-rating song found so far.
Now a song's rating is set to 0 by default, so when a rating for that song is
not found, it will be 0.