Write playlist to an m3u file.
Created a new method, write_playlist_to_m3u_file, which takes a vector of Song
objects (the playlist) and writes it to an m3u file.  At the moment, this file
is very basic, as it does not include any extended metadata.  This file can be
successfully read by VLC.