Find the "Music" playlist after parsing the library.
In order to pick music from only one playlist, we need to find the playlist
after parsing the library.  For now, we are hard-coding using the "Music"
playlist.  After parsing the songs and playlists, find the "Music" playlist
and store it for later use.  If this playlist cannot be found, return with a
non-zero error code, as there is nothing else we can do without that playlist.

Also, print out a message after successfully finding the playlist with the
name of the playlist that will be used and the number of songs in that
playlist.