Referenced the boost program_options dll when building.
This version of the file is very messy, as I was trying to figure out how to
include a boost library, specifically one needed to be compiled.  The final
version does work, but does not use a compiler switch (such as I, L, or l) to
reference where the dll is.  At a later point I will revisit this and figure
out how to set this up better.

Also, the boost libraries and references will not (at this point, at least) be
added to the repository, as they are many in number and size.  I need to look
into how to use the bcp program to pull out just the files I need to compile
this program.