Added method to parse command line arguments.
Added method parse_command_line, which takes as parameters the command line
arguments to the program, and returns a boost::program_options::variables_map
object, which has all of the parsed arguments.  For now, since the
options_description object lives just inside this method, if the "help"
parameter is found, print out the usage information here.