Use values from command line parameters for settings.
Use the values passed in at the command line (or default values) for the name
of the output playlist file, name of the iTunes playlist name to get tracks
from, and the file name to use for the library input.  Also, pass the output
playlist file name to the write_playlist_to_m3u_file, so it can be used when
opening the file.