Added parameter for rating weighting scheme.
The rating weighting scheme will determine how to weight tracks with different
weights relative to each other.  The default is to weight them linearly; this
is used if no scheme is specified.  If an invalid scheme is specified, the
program quits after providing an error message.  The valid schemes right now
are 'linear' and 'iTunes'.