Do not compile gcc and c++ libraries dynamically.
The boost program_options library appears to not have the gcc and c++
libraries statically linked, so the resulting executable needs these in the
same directory with it.  I have not yet found out how to compile those to
statically.  Since these need to be in the output directory anyway, now I have
switched compiling savant.exe to dyanically reference them.  As a result of
all these, these two libraries need to be in the output directory with the
exe.